+++ src/main/java/com/api/identity/enums/InvitationStatus.java
@@ -3,5 +3,6 @@ package com.api.identity.enums;
 public enum InvitationStatus {
     PENDING,
     ACCEPTED,
-    REJECTED
+    REJECTED,
+    CANCELLED
 }